travel-project/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
indic-nlp-library==0.92
iNLTK==0.8
# spacy==3.7.2
# transformers==4.35.2
# torch==2.1.1
nltk==3.8.1
pandas==2.1.3
googletrans==4.0.0rc1